rename proxy services
